requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
qwen-vl-utils>=0.0.3
opencv-python>=4.8.0
Pillow>=10.0.0
tokenizers>=0.14.0
nltk>=3.8.1
rouge-score>=0.1.2
sentence-transformers>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
datasets>=2.14.0
accelerate>=0.21.0
peft>=0.5.0
pycocotools>=2.0.7
bleu>=0.3
cider>=1.0
requests>=2.31.0
openai>=1.0.0
pinecone-client>=2.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
jsonlines>=3.1.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.2.0
deepspeed>=0.10.0
